Have transformation animations have end keyframe offset by 10
For new users, I think it is hard to use most animations, because the start and end keyframe node markers by default are placed at the same location AND BOTH NODE MARKERS ARE SELECTED WHEN YOU CLICK ON ONE.  It took me a long time to figure out I have to manually offset a keyframe X parameter to see and manipulate node markers individually.  Starting off with star and end keyframe node markers in different locations helps visualize and use the tool.  The arbitrary value of 10 shouldn't be an issue, since the user will probably be going to transform the markers around anyways.

There is still a problem that if a keyframe is duplicated, that the duplicate node marker  will be in the same location (and both original and duplicate node marker will be selected).  Eventually a feature should be made to optionally select only the top node marker.  How that should work is questionable, since at times having node markers at the same location be both selected is beneficial (e.g. rotation pivot point translated for both start and end keyframes at the same time)